Inline the `intersperse` function.
I didn't like the name of the file and it was only used in one place.
This commit also makes the toml formatting consistent in the `netlify.toml` file.
